+++ starkbank/__init__.py
@@ -106,6 +106,12 @@ from .paymentrequest.__paymentrequest import PaymentRequest
 from . import invoice
 from .invoice.__invoice import Invoice
 
+from . import invoicepullrequest
+from .invoicepullrequest.__invoicepullrequest import InvoicePullRequest
+
+from . import invoicepullsubscription
+from .invoicepullsubscription.__invoicepullsubscription import InvoicePullSubscription
+
 from . import dictkey
 from .dictkey.__dictkey import DictKey